책 등록, 수정, 삭제 시 유저 확인 로직 추가

+++ bookforeast/src/main/java/com/project/bookforeast/library/entity/Library.java
@@ -30,7 +30,7 @@ public class Library {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long libraryId;
-	private int upperLibraryId;
+	private Long upperLibraryId;
 	private String libraryName;
 	
 	@CreationTimestamp
